feat: IssueMapper 생성
