+++ backend/src/main/java/kr/codesquad/issuetracker/infrastructure/persistence/mapper/LabelSimpleMapper.java
@@ -1,0 +1,22 @@
+package kr.codesquad.issuetracker.infrastructure.persistence.mapper;
+
+import org.springframework.util.StringUtils;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class LabelSimpleMapper {
+	private String name;
+	private String fontColor;
+	private String backgroundColor;
+
+	public boolean hasValue() {
+		return StringUtils.hasText(name);
+	}
+}
